requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
coverage==7.5.3
drf-yasg==1.21.7
setuptools==70.1.0
stripe==10.0.0
